♻️  move sample_entropy function to stats.py
🐛  remove import of sampen module from series init file
