+++ tinyshift/series/__init__.py
@@ -6,4 +6,3 @@
 from .outlier import *
 from .forecastability import *
 from .stats import *
-from .sampen import *
