+++ tinyshift/series/stats.py
@@ -150,6 +150,89 @@ def relative_strength_index(
     return rsi
 
 
+def sample_entropy(
+    X: Union[np.ndarray, List[float]],
+    m: int = 1,
+    tolerance: float = None,
+) -> np.ndarray:
+    """
+    Compute the Sample Entropy (SampEn) of a 1D time series.
+    Sample Entropy is a measure of complexity or irregularity in a time series.
+    It quantifies the likelihood that similar patterns in the data will not be followed by additional similar patterns.
+    Parameters
+    ----------
+    X : array-like, shape (n_samples,)
+        1D time series data.
+    m : int
+        Length of sequences to be compared (embedding dimension).
+    tolerance : float, optional (default=None)
+        Tolerance for accepting matches. If None, it is set to 0.1 * std(X).
+    Returns
+    -------
+    sampen : float
+        The Sample Entropy of the time series. Returns np.nan if A or B is zero.
+    References
+    ----------
+    - Richman, J. S., & Moorman, J. R. (2000). Physiological time-series analysis using approximate entropy and sample entropy. American Journal of Physiology-Heart and Circulatory Physiology, 278(6), H2039-H2049.
+    - Lake, D. E., Richman, J. S., Griffin, M. P., & Moorman, J. R. (2002). Sample entropy analysis of neonatal heart rate variability. American Journal of Physiology-Regulatory, Integrative and Comparative Physiology, 283(3), R789-R797.
+    Notes
+    -----
+    - SampEn is less biased than Approximate Entropy (ApEn) and does not count self-matches.
+    - Higher SampEn values indicate more complexity and irregularity in the time series.
+    - The function assumes the input time series is 1-dimensional.
+    - The function uses the Chebyshev distance (maximum norm) for comparing sequences.
+    - If either A or B is zero, SampEn is undefined and np.nan is returned.
+    """
+
+    X = np.asarray(X, dtype=np.float64)
+
+    if X.ndim != 1:
+        raise ValueError("Input data must be 1-dimensional")
+
+    if tolerance is None:
+        tolerance = 0.1 * np.std(X)
+
+    n = len(X)
+
+    Xm = np.array([X[i : i + m] for i in range(n - m)])
+
+    Xm1 = np.array([X[i : i + m + 1] for i in range(n - m - 1)])
+
+    def count_matches(X_templates, tol):
+        """
+        Count the number of matching template pairs within the given tolerance. Chebyshev distance is used.
+        Parameters
+        ----------
+        X_templates : ndarray, shape (N, m) or (N, m+1)
+            Array of template vectors.
+        tol : float
+            Tolerance for accepting matches.
+        Returns
+        -------
+        count : int
+            Number of matching template pairs.
+        """
+
+        count = 0
+        N = len(X_templates)
+        for i in range(N):
+            diff = np.abs(X_templates[i] - X_templates[i + 1 :])
+            max_diff = np.max(diff, axis=1)
+            count += np.sum(max_diff < tol)
+        return count
+
+    B = count_matches(Xm, tolerance)
+
+    A = count_matches(Xm1, tolerance)
+
+    if A > 0 and B > 0:
+        sampen = -np.log(A / B)
+    else:
+        sampen = np.nan
+
+    return sampen
+
+
 def entropy_volatility(
     X: Union[np.ndarray, List[float]],
     rolling_window: int = 60,
